chore: cleanup unused types and move anys to src/types
