+++ src/cdk/v2/utils.ts
@@ -5,7 +5,7 @@ import { PlatformError } from '@rudderstack/integrations-lib';
 import logger from '../../logger';
 import { generateErrorObject } from '../../v0/util';
 import tags from '../../v0/util/tags';
-import { CatchErr } from '../../util/types';
+import { CatchErr } from '../../types';
 
 const CDK_V2_ROOT_DIR = __dirname;
